feat: Add account creation, login, and password modification features
- Implemented EnumsCreateAccountWidget for user type selection (client/female driver).
- Created RichTextCreateAnAccountWidget for account creation prompts.
- Developed ForgetPasswordPage for password recovery functionality.
- Added LoginPage with email and password fields, including validation.
- Introduced RichTextLoginWidget for login prompts and account creation navigation.
- Implemented ModifyPasswordPage for changing user passwords.
- Created OnboardingPage with a PageView for onboarding experience.
- Added Dot widget for onboarding page indicators.
- Developed SplashPage for initial app loading screen.
- Implemented VerificationCodePage for code verification during password recovery.
- Added TextRichReceiveTheCode widget for resending verification codes.
- Created WelcomePage for post-login navigation.
- Updated localization files for new strings and translations.
- Removed unused provider file.
- Updated dependencies in pubspec.lock for improved functionality and compatibility.
