+++ lib/core/components/custom_icon_change_language.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import '../constant/color_manger.dart';
+
+import '../constant/app_colors.dart';
 import '../router/router.dart';
 
 @RoutePage()
@@ -17,7 +18,7 @@ class CustomIconChangeLanguagePage extends StatelessWidget {
       },
       icon: const Icon(
         Icons.language,
-        color: ColorManger.kPrimaryColor,
+        color: AppColors.kPrimaryColor,
       ),
     );
   }
