+++ lib/core/constant/app_colors.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const kPrimaryColor = Color(0xff805997);
+  static const kWhite = Colors.white;
+  static const kBorderColor = Color(0xffbab8b9);
+}
