+++ lib/core/constant/app_images.dart
@@ -1,0 +1,7 @@
+class AppImages {
+  static const kOndoarding1 = 'assets/image/car5_1.png';
+  static const kOndoarding2 = 'assets/image/car2_1.png';
+  static const kOndoarding3 = 'assets/image/car_6.png';
+  static const kForgetpassword = 'assets/image/forget_password_1.png';
+  static const kNoprofile = 'assets/image/no-profile.jpg';
+}
