--- lib/core/constant/color_manger.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorManger {
-  static const kPrimaryColor = Color(0xff805997);
-  static const kWhite = Colors.white;
-  static const kBorderColor = Color(0xffbab8b9);
-}
