--- lib/core/constant/svg_manger.dart
@@ -1,50 +1,0 @@
-class SvgManger {
-  static const kAlertCircle = 'assets/svg/alert-circle.svg';
-  static const kArrowleft = 'assets/svg/arrow-left.svg';
-  static const kArrowright = 'assets/svg/arrow-right.svg';
-  static const kBelloff = 'assets/svg/bell-off.svg';
-  static const kBell = 'assets/svg/bell.svg';
-  static const kCamera = 'assets/svg/camera.svg';
-  static const kCheck = 'assets/svg/check.svg';
-  static const kChevrondown = 'assets/svg/chevron-down.svg';
-  static const kChevronleft = 'assets/svg/chevron-left.svg';
-  static const kChevronright = 'assets/svg/chevron-right.svg';
-  static const kChevronup = 'assets/svg/chevron-up.svg';
-  static const kClock = 'assets/svg/clock.svg';
-  static const kCredit = 'assets/svg/credit.svg';
-  static const kCurrentlocation = 'assets/svg/currentlocation.svg';
-  static const kDelete = 'assets/svg/delete.svg';
-  static const kEdit2 = 'assets/svg/edit-2.svg';
-  static const kEdit = 'assets/svg/edit.svg';
-  static const kEyeoff = 'assets/svg/eye-off.svg';
-  static const kEye = 'assets/svg/eye.svg';
-  static const kHeart = 'assets/svg/heart.svg';
-  static const kHome = 'assets/svg/home.svg';
-  static const kLocation = 'assets/svg/location.svg';
-  static const kLock = 'assets/svg/lock.svg';
-  static const kLogin = 'assets/svg/log-in.svg';
-  static const kLogout = 'assets/svg/log-out.svg';
-  static const kLogo = 'assets/svg/logo.svg';
-  static const kMail = 'assets/svg/mail.svg';
-  static const kMappin = 'assets/svg/map-pin.svg';
-  static const kMenu = 'assets/svg/menu.svg';
-  static const kMessagecircle = 'assets/svg/message-circle.svg';
-  static const kMessagesquare = 'assets/svg/message-square.svg';
-  static const kMicoff = 'assets/svg/mic-off.svg';
-  static const kMic = 'assets/svg/mic.svg';
-  static const kMorehorizontal = 'assets/svg/more-horizontal.svg';
-  static const kMorevertical = 'assets/svg/more-vertical.svg';
-  static const kPhone = 'assets/svg/phone.svg';
-  static const kSearch = 'assets/svg/search.svg';
-  static const kSettings = 'assets/svg/settings.svg';
-  static const kShare2 = 'assets/svg/share-2.svg';
-  static const kSliders = 'assets/svg/sliders.svg';
-  static const kStar = 'assets/svg/star.svg';
-  static const kThumbsdown = 'assets/svg/thumbs-down.svg';
-  static const kThumbsup = 'assets/svg/thumbs-up.svg';
-  static const kTruck = 'assets/svg/truck.svg';
-  static const kUser = 'assets/svg/user.svg';
-  static const kXoctagon = 'assets/svg/x-octagon.svg';
-  static const kXsquare = 'assets/svg/x-square.svg';
-  static const kX = 'assets/svg/x.svg';
-}
