--- lib/core/models/enums_local_language.dart
@@ -1,1 +1,0 @@
-enum EnumChangeLocalLanguage { intialStat, arabicLanguage, englishLanguage }
