+++ lib/core/router/router.dart
@@ -1,15 +1,15 @@
 import 'package:auto_route/auto_route.dart';
 import '../components/custom_icon_change_language.dart';
 
-import '../../home/pages/change_language/change_language_page.dart';
-import '../../home/pages/create_an_account/create_an_account_page.dart';
-import '../../home/pages/forget_password/forget_password_page.dart';
-import '../../home/pages/login/login_page.dart';
-import '../../home/pages/modify_password/modify_password_page.dart';
-import '../../home/pages/on_boarding/onboarding_page.dart';
-import '../../home/pages/splash/splash_page.dart';
-import '../../home/pages/verification_code/verification_code_page.dart';
-import '../../home/pages/welcome/welcome_page.dart';
+import '../../features/presentation/change_language/change_language_page.dart';
+import '../../features/presentation/create_an_account/view/create_an_account_page.dart';
+import '../../features/presentation/forget_password/forget_password_page.dart';
+import '../../features/presentation/login/view/login_page.dart';
+import '../../features/presentation/modify_password/modify_password_page.dart';
+import '../../features/presentation/on_boarding/view/onboarding_page.dart';
+import '../../features/presentation/splash/splash_page.dart';
+import '../../features/presentation/verification_code/view/verification_code_page.dart';
+import '../../features/presentation/welcome/welcome_page.dart';
 
 part 'router.gr.dart';
