+++ lib/features/data/cubits/app_change_theme_cubit/app_change_theme_cubit.dart
@@ -1,0 +1,7 @@
+import 'package:bloc/bloc.dart';
+
+part 'app_change_theme_state.dart';
+
+class AppChangeThemeCubit extends Cubit<AppChangeThemeState> {
+  AppChangeThemeCubit() : super(AppChangeThemeInitial());
+}
