+++ lib/features/data/cubits/app_change_theme_cubit/app_change_theme_state.dart
@@ -1,0 +1,6 @@
+part of 'app_change_theme_cubit.dart';
+
+
+abstract class AppChangeThemeState {}
+
+final class AppChangeThemeInitial extends AppChangeThemeState {}
