+++ lib/features/data/models/enums_local_language.dart
@@ -1,0 +1,1 @@
+enum Localization { intial, arabic, english }
