--- lib/home/data/manger/app_language/app_language_state.dart
@@ -1,12 +1,0 @@
-part of 'app_language_cubit.dart';
-
-@immutable
-sealed class AppLanguageState {}
-
-final class AppLanguageInitial extends AppLanguageState {}
-
-final class AppLanguageChange extends AppLanguageState {
-  final String isChange;
-
-  AppLanguageChange({required this.isChange});
-}
