--- lib/provider/provider.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-final themeProvider = StateProvider<bool>((ref) {
-  return true;
-});
